added tailwindcss config and restructure the project.
