+++ src/App.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
-import Navbar from './Navbar';
-import Landingpage from './Landingpage';
-import './App.css';
-import './landingpage.css';
+import Navbar from './components/Navbar';
+import Landingpage from './components/Landingpage';
+
 
 
 // import { render } from '@testing-library/react';
@@ -18,9 +17,7 @@ import './landingpage.css';
 function App() {
     return (
       <div className="App">
-        {/* <register /> */}
-        {/* <h1 > landiFFFngpage</h1> */}
-        <p>hwllo there iam typing with only one hand and fast</p>
+
         <Navbar />
         <Landingpage />
