+++ src/components/Navbar.js
@@ -1,0 +1,17 @@
+import React from 'react'  //utilizing useState
+
+
+function   Navbar() {
+
+
+    return (
+        <>
+       <nav className="navbar">
+
+       </nav> 
+       
+        </>
+    )
+}
+
+export default Navbar;
